+++ mangadex_downloader/format/comic_book.py
@@ -103,7 +103,8 @@ class ComicBookArchive(BaseFormat):
             desc='cbz_progress',
             total=len(images),
             initial=0,
-            unit='item'
+            unit='item',
+            disable=not self.progress_bar
         )
 
         for im_path in images: